requirements.txt
Django==4.2.19
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0

